multi_agent_demo/requirements.txt
# multi_agent_demo/requirements.txt
langchain==0.1.13
llama-index==0.10.14
chromadb==0.4.22
openai==1.3.9
tiktoken==0.5.1
python-dotenv==1.0.1
requests==2.31.0
httpx==0.24.1
pydantic==1.10.13
typing-extensions==4.6.3
nest-asyncio==1.6.0
fastapi==0.95.2
uvicorn[standard]==0.22.0